removing old precog files
