+++ mapbox/app/app/templates/index.html
@@ -302,7 +302,7 @@
 				var select = document.getElementById("precog-select");
 				for (var i = 0; i < data.children.length; i++) {
 					select.options[select.options.length] = new Option(data.children[i].type, data.children[i].id);
-					console.log(data[i].type);
+					// console.log(data.children[i].type);
 				}
 			});
